Move legacy implementations to Core/Legacy
