+++ Core/Legacy/SpecularMagneticStrategy_v1.cpp
