+++ Core/Legacy/SpecularMagneticStrategy_v2.cpp
