+++ Tests/UnitTests/Core/Legacy/MatrixRTCoefficients_v1Test.cpp
@@ -1,4 +1,4 @@
-#include "Tests/UnitTests/Core/Legacy/MatrixRTCoefficients_v1.h"
+#include "Core/Legacy/MatrixRTCoefficients_v1.h"
 #include "Tests/GTestWrapper/google_test.h"
 
 class MatrixRTCoefficients_v1Test : public ::testing::Test {
